chore: remove legacy Transxchange publisher code
