+++ app/api/module/Api/src/Domain/CommandHandler/Bus/Ebsr/RequestMapQueue.php
@@ -56,7 +56,6 @@ final class RequestMapQueue extends AbstractCommandHandler implements AuthAwareI
         $optionData = [
             'scale' => $command->getScale(),
             'id' => $entityId,
-            'fromNewEbsr' => $command->getFromNewEbsr(),
             'regNo' => $busReg->getRegNo(),
             'licence' => $busReg->getLicence()->getId(),
             'user' => $this->getCurrentUser()->getId()
