+++ app/api/module/Api/src/Entity/System/FeatureToggle.php
@@ -25,7 +25,6 @@ class FeatureToggle extends AbstractFeatureToggle
     public const SELFSERVE_SURRENDER = 'ss_surrender';
     public const BACKEND_SURRENDER = 'back_surrender';
     public const MESSAGING = 'messaging';
-    public const BACKEND_TRANSXCHANGE = 'transxchange_connection';
     public const USE_NEW_ADDRESS_SERVICE = 'use_new_address_service';
     public const TRANSPORT_CONSULTANT_ROLE = 'transport_consultant_role';
