+++ app/api/test/module/Api/src/Domain/Command/Bus/Ebsr/ProcessRequestMapTest.php
@@ -11,14 +11,13 @@ use Dvsa\Olcs\Api\Domain\Command\Bus\Ebsr\ProcessRequestMap;
  */
 class ProcessRequestMapTest extends \PHPUnit\Framework\TestCase
 {
-    public function testStructure()
+    public function testStructure(): void
     {
         $id = 1;
         $user = 2;
         $scale = 'small';
         $licence = 'licence';
         $regNo = '123/45678';
-        $fromNewEbsr = true;
 
         $command = ProcessRequestMap::create(
             [
@@ -26,8 +25,7 @@ class ProcessRequestMapTest extends \PHPUnit\Framework\TestCase
                 'user' => $user,
                 'scale' => $scale,
                 'licence' => $licence,
-                'regNo' => $regNo,
-                'fromNewEbsr' => $fromNewEbsr
+                'regNo' => $regNo
             ]
         );
 
@@ -36,6 +34,5 @@ class ProcessRequestMapTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($scale, $command->getScale());
         $this->assertEquals($licence, $command->getLicence());
         $this->assertEquals($regNo, $command->getRegNo());
-        $this->assertEquals($fromNewEbsr, $command->getFromNewEbsr());
     }
 }
