+++ app/api/test/module/Api/src/Service/Ebsr/TransExchangeClientFactoryTest.php
@@ -9,7 +9,6 @@ use Dvsa\Olcs\Api\Service\AccessToken\Provider;
 use Dvsa\Olcs\Api\Service\Ebsr\TransExchangeClient;
 use Dvsa\Olcs\Api\Service\Ebsr\TransExchangeClientFactory;
 use Psr\Container\ContainerInterface;
-use Dvsa\Olcs\Api\Service\Toggle\ToggleService;
 use Mockery\Adapter\Phpunit\MockeryTestCase as TestCase;
 use Mockery as m;
 use Olcs\XmlTools\Filter\MapXmlFile;
@@ -32,55 +31,6 @@ class TransExchangeClientFactoryTest extends TestCase
         $sut->__invoke($mockSl, TransExchangeClient::class);
     }
 
-    public function testInvoke(): void
-    {
-        $oAuthOptions = ['options'];
-        $config = [
-            'transexchange_publisher' => [
-                'uri' => 'http://localhost:8080/txc/',
-                'options' => ['argseparator' => '~'],
-                'template_file' => 'template.xml',
-                'oauth2' => $oAuthOptions,
-            ]
-        ];
-
-        $mockSpec = m::mock(SpecificationInterface::class);
-        $mockFilter = m::mock(MapXmlFile::class);
-        $mockFilter->shouldReceive('setMapping')
-            ->once()
-            ->with($mockSpec);
-
-        $mockParser = m::mock(ParseXmlString::class);
-
-        $mockXsdValidator = m::mock(Xsd::class);
-        $mockXsdValidator->shouldReceive('setXsd')
-            ->once()
-            ->with(TransExchangeClientFactory::PUBLISH_XSD);
-
-        $mockTokenProvider = m::mock(Provider::class);
-        $mockTokenProvider->expects('getToken')->withNoArgs()->andReturn('token');
-
-        $mockToggle = m::mock(ToggleService::class);
-        $mockToggle->shouldReceive('getToggleService')->andReturn($mockToggle);
-        $mockToggle->shouldReceive('isEnabled')->with(FeatureToggle::BACKEND_TRANSXCHANGE)->andReturn(false);
-        $mockSl = m::mock(ContainerInterface::class);
-        $mockSl->shouldReceive('get')->with('config')->andReturn(['ebsr' => $config]);
-        $mockSl->shouldReceive('get')->with('FilterManager')->andReturnSelf();
-        $mockSl->shouldReceive('get')->with('ValidatorManager')->andReturnSelf();
-        $mockSl->shouldReceive('get')->with('TransExchangePublisherXmlMapping')->andReturn($mockSpec);
-        $mockSl->shouldReceive('get')->with(MapXmlFile::class)->andReturn($mockFilter);
-        $mockSl->shouldReceive('get')->with(ParseXmlString::class)->andReturn($mockParser);
-        $mockSl->shouldReceive('get')->with(Xsd::class)->andReturn($mockXsdValidator);
-        $mockSl->shouldReceive('get')->with(ToggleService::class)->andReturn($mockToggle);
-        $mockSl->shouldReceive('build')->with(Provider::class, $oAuthOptions)->andReturn($mockTokenProvider);
-
-        $sut = new TransExchangeClientFactory();
-        $service = $sut->__invoke($mockSl, TransExchangeClient::class);
-
-        $this->assertInstanceOf(TransExchangeClient::class, $service);
-    }
-
-
     /**
      * @throws ContainerExceptionInterface
      * @throws NotFoundExceptionInterface
@@ -91,7 +41,6 @@ class TransExchangeClientFactoryTest extends TestCase
         $config = [
             'ebsr' => [
                 'transexchange_publisher' => [
-                    'uri' => 'http://localhost:8080/txc/',
                     'new_uri' => 'http://txc.dev/',
                     'options' => ['argseparator' => '~'],
                     'template_file' => 'template.xml',
@@ -99,9 +48,6 @@ class TransExchangeClientFactoryTest extends TestCase
                 ],
             ],
         ];
-        $mockToggle = m::mock(ToggleService::class);
-        $mockToggle->shouldReceive('getToggleService')->andReturn($mockToggle);
-        $mockToggle->shouldReceive('isEnabled')->with(FeatureToggle::BACKEND_TRANSXCHANGE)->andReturn(true);
 
         $mockSpec = m::mock(SpecificationInterface::class);
         $mockFilter = m::mock(MapXmlFile::class);
@@ -126,7 +72,6 @@ class TransExchangeClientFactoryTest extends TestCase
         $mockSl->shouldReceive('get')->with(MapXmlFile::class)->andReturn($mockFilter);
         $mockSl->shouldReceive('get')->with(ParseXmlString::class)->andReturn($mockParser);
         $mockSl->shouldReceive('get')->with(Xsd::class)->andReturn($mockXsdValidator);
-        $mockSl->shouldReceive('get')->with(ToggleService::class)->andReturn($mockToggle);
         $mockSl->shouldReceive('build')->with(Provider::class, $oAuthOptions)->andReturn($mockTokenProvider);
 
         $sut = new TransExchangeClientFactory();
